+++ src/datasets/index.ts
@@ -163,7 +163,6 @@ export {
 export { DatasetLinkedCollection } from './domain/models/DatasetLinkedCollection'
 export { DatasetType } from './domain/models/DatasetType'
 export { DatasetTypeDTO } from './domain/dtos/DatasetTypeDTO'
-export { StorageDriver } from './domain/models/StorageDriver'
 export { DatasetUploadLimits } from './domain/models/DatasetUploadLimits'
 export {
   FileTreeNode,